worker-node/requirements.txt
flask==3.1.1
pymongo==4.13.2
pika==1.3.2
python-dotenv==1.1.0
python-json-logger==3.3.0